# Improve performance for #one? and #many?