@byroot 2021-05-05
I can't put my finger on it, but I feel like I heard about this idea not so long ago.

Either way it sounds good to me. Could you please clean your commits? There's an extra merge commit that shouldn't be there.