@eugeneius 2021-05-05
https://github.com/rails/rails/pull/27175 was the previous attempt at this. Matthew's comment about the result no longer being memoised is still valid here.