@byroot 2021-05-05
Right, at the very least that "partial count" should be memoized as to make sure subsequence calls don't trigger a query.