@gonzaloriestra 2021-05-05
Thanks for the comments! I didn't know about that attempt, sorry. I had a look at previous issues/PRs, but there's too much content here... :sweat_smile: 

Regarding the memoization, the query is avoided when it's already loaded (see [this line](https://github.com/rails/rails/pull/42143/files#diff-18a561656864ea240daf46bdb1f50faace49f9ef74b90bcf667d9bbb17fce084R297)). I think it's covered by [this test](https://github.com/rails/rails/blob/70c5496542e5dc82ca28840cb01e710200ce5d14/activerecord/test/cases/associations/has_many_associations_test.rb#L2351). Or am I missing something?