@byroot 2021-05-06
> Regarding the memoization, the query is avoided when it's already loaded (see [this line](https://github.com/rails/rails/pull/42143/files#diff-18a561656864ea240daf46bdb1f50faace49f9ef74b90bcf667d9bbb17fce084R297)). I think it's covered by [this test](https://github.com/rails/rails/blob/70c5496542e5dc82ca28840cb01e710200ce5d14/activerecord/test/cases/associations/has_many_associations_test.rb#L2351). Or am I missing something?

Yes, the memoization we are talking about is: `3.times { scope.one? }`, which would trigger 3 queries.

I think the result of `limit(2).count` could be memoized and re-used. It could even stored as the `count` memoization if it returns `1` or `0`. But maybe that's too brittle.