@gonzaloriestra 2021-05-06
But that was also triggering 3 queries before, wasn't it? And even with [size](https://github.com/rails/rails/blob/70c5496542e5dc82ca28840cb01e710200ce5d14/activerecord/lib/active_record/relation.rb#L265)... :thinking: 

Anyway, I agree. I've finally memoized the `limit(2).count` to share it between both methods. What do you think?