@byroot 2021-05-06
Hum, my bad. I assumed `size` was memoized as well, but it seems I was wrong. That's what Matthew said in https://github.com/rails/rails/pull/27175#issuecomment-431271531, did that change recently?