@gonzaloriestra 2021-05-06
I think it hasn't changed, it was similar in that moment: https://github.com/marshall-lee/rails/blob/046c3b6edf28d9604eb99b09bfb9501d5c50eee1/activerecord/lib/active_record/relation.rb#L240