@byroot 2021-05-06
Hum, then cc @matthewd, maybe you remember what you ment by "memoization" at the time?