@gonzaloriestra 2021-05-13
To ensure this PR doesn't fall into oblivion like the last attempt, what about merging the changes without memoization (as it was before), and opening a new issue to discuss adding it to all the convenient places in `Relation` (and not only to `one?` and `many?`)?