@byroot 2021-05-13
Ok, I looked at the code at the time of @matthewd's comment, and I'm pretty sure `#size` wasn't memoizing:

```
    # Returns size of the records.
    def size
      loaded? ? @records.length : count(:all)
    end
```

So I think this PR isn't degrading any performance in existing scenarios. So I'm 👍 . I'll merge once the method is made private and the merge conflict resolved.