@byroot 2021-05-13
Wait, I'm having a doubt here, does this even work?  I'm more familiar with MySQL, so maybe it works with postgres, but `SELECT COUNT(*) LIMIT 2` will limit the number of returned rows, not the number of counted rows, right?

That's why the old PR is much more complicated, it does `SELECT 1 AS one LIMIT` and check the number of returned rows.