@gonzaloriestra 2021-05-13
Right, `SELECT COUNT(*) LIMIT 2` doesn't limit the scanned rows (at least in MySQL and PostgreSQL). I wrote that simplified query in the description, but it's not accurate (I've updated it).

The generated query is not that one. `Post.limit(2).count` is running the complex query in fact: `SELECT COUNT(count_column) FROM (SELECT 1 AS count_column FROM `posts` LIMIT 2) subquery_for_count`.

If you have a look at the benchmark, I think it's quite clear that it does work.