@byroot 2021-05-13
Ok, nevermind you are right. `limit(2).count` will generate a subquery: `SELECT COUNT(*) FROM (SELECT 1 AS one FROM "posts" LIMIT 2) subquery_for_count`. I just tested it.

My bad.