@matthewd 2021-05-14
Sorry I missed this earlier -- yeah, it seems I was confidently incorrect about `size` memoizing :confused: